Add JsonBlobParser and refactor related utilities
Created JsonBlobParser.java for JSON parsing and refactored ConceptResultSetUtil to use it for clearer separation of concerns.
