+++ src/main/java/edu/harvard/dbmi/avillach/dictionary/concept/ConceptResultSetUtil.java
@@ -2,29 +2,31 @@ package edu.harvard.dbmi.avillach.dictionary.concept;
 
 import edu.harvard.dbmi.avillach.dictionary.concept.model.CategoricalConcept;
 import edu.harvard.dbmi.avillach.dictionary.concept.model.ContinuousConcept;
-import org.json.JSONArray;
-import org.json.JSONException;
+import edu.harvard.dbmi.avillach.dictionary.util.JsonBlobParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Component
 public class ConceptResultSetUtil {
 
     private static final Logger log = LoggerFactory.getLogger(ConceptResultSetUtil.class);
+    private final JsonBlobParser jsonBlobParser;
+
+    @Autowired
+    public ConceptResultSetUtil(JsonBlobParser jsonBlobParser) {
+        this.jsonBlobParser = jsonBlobParser;
+    }
 
     public CategoricalConcept mapCategorical(ResultSet rs) throws SQLException {
         return new CategoricalConcept(
             rs.getString("concept_path"), rs.getString("name"), rs.getString("display"), rs.getString("dataset"),
-            rs.getString("description"), rs.getString("values") == null ? List.of() : parseValues(rs.getString("values")),
+            rs.getString("description"), rs.getString("values") == null ? List.of() : jsonBlobParser.parseValues(rs.getString("values")),
             rs.getBoolean("allowFiltering"), rs.getString("studyAcronym"), null, null
         );
     }
@@ -32,53 +34,11 @@ public class ConceptResultSetUtil {
     public ContinuousConcept mapContinuous(ResultSet rs) throws SQLException {
         return new ContinuousConcept(
             rs.getString("concept_path"), rs.getString("name"), rs.getString("display"), rs.getString("dataset"),
-            rs.getString("description"), rs.getBoolean("allowFiltering"), parseMin(rs.getString("values")),
-            parseMax(rs.getString("values")), rs.getString("studyAcronym"), null
+            rs.getString("description"), rs.getBoolean("allowFiltering"), jsonBlobParser.parseMin(rs.getString("values")),
+            jsonBlobParser.parseMax(rs.getString("values")), rs.getString("studyAcronym"), null
         );
     }
 
-    public List<String> parseValues(String valuesArr) {
-        try {
-            ArrayList<String> vals = new ArrayList<>();
-            JSONArray arr = new JSONArray(valuesArr);
-            for (int i = 0; i < arr.length(); i++) {
-                vals.add(arr.getString(i));
-            }
-            return vals;
-        } catch (JSONException ex) {
-            return List.of();
-        }
-    }
 
-    public Float parseMin(String valuesArr) {
-        return parseFromIndex(valuesArr, 0);
-    }
 
-    private Float parseFromIndex(String valuesArr, int index) {
-        try {
-            JSONArray arr = new JSONArray(valuesArr);
-            if (arr.length() != 2) {
-                return 0F;
-            }
-            Object raw = arr.get(index);
-            return switch (raw) {
-                case Double d -> d.floatValue();
-                case Integer i -> i.floatValue();
-                case String s -> Double.valueOf(s).floatValue();
-                case BigDecimal d -> d.floatValue();
-                case BigInteger i -> i.floatValue();
-                default -> 0f;
-            };
-        } catch (JSONException ex) {
-            log.warn("Invalid json array for values: ", ex);
-            return 0F;
-        } catch (NumberFormatException ex) {
-            log.warn("Valid json array but invalid val within: ", ex);
-            return 0F;
-        }
-    }
-
-    public Float parseMax(String valuesArr) {
-        return parseFromIndex(valuesArr, 1);
-    }
 }
